requirements.txt
aiogram==3.10.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.1.5